Under-relax fine grid smoother for enthalpy and add more output
green-3-20km_ent_muk_wdg_np8 failed over the weekend for an unknown reason.
It was not caused by the change to point relaxation because it worked saturday
but not sunday. Changes to trilinos/albany were minimal and I doubt that
caused the issue. Under-relaxing the fine grid smoother seemed to fix the
issue. I tried 0.9, 0.95 and 0.99 - 0.95 seemed to work the best.